requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
cfgv==3.4.0
channels==4.2.0
channels_redis==4.2.1
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
distlib==0.3.9
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
filelock==3.16.1
hyperlink==21.0.0
identify==2.6.3
idna==3.10
incremental==24.7.2
inflection==0.5.1
joblib==1.4.2
msgpack==1.1.0
nodeenv==1.9.1
numpy==2.1.3
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
pre_commit==4.0.1
psycopg==3.2.3
psycopg-binary==3.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
scikit-learn==1.5.2
scipy==1.14.1
service-identity==24.2.0
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.2
threadpoolctl==3.5.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
virtualenv==20.28.0
zope.interface==7.2
